MINI_PROJECT/rag-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
langchain>=0.0.335
langchain-community>=0.0.13
langchain-core>=0.1.12
langchain-pinecone>=0.1.0
langchain-google-genai==0.0.5
google-generativeai==0.3.1
pinecone>=6.0.0
tiktoken==0.5.1
redis==5.0.1
loguru==0.7.2
pytz==2023.3
requests==2.31.0
httpx==0.25.0
numpy==1.24.3
pandas==2.0.3
tenacity==8.2.3
aiohttp==3.8.6
jinja2==3.1.2
python-multipart==0.0.6
# Added for MongoDB
motor==3.3.2
pymongo==4.6.1
# Added for HuggingFace embeddings
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2
